requirements.txt
kafka-python
pyspark
fastapi
uvicorn
openai
pandas
streamlit
sqlite3
sqlalchemy
